Ben The Gamer Problem
Ben is one of the best gamers in India. He also happens to be an excellent programmer. 
So, he likes to play games which require use of both gaming skills as well as programming skills. 
One such game is SpaceWar.

In this game there are N levels and M types of available weapons. 
The levels are numbered from 0 to N-1 and the weapons are numbered from 0 to M-1 . 
Ben can clear these levels in any order. 
In each level, some subset of these M weapons is required to clear this level. 
If in a particular level, Ben needs to buy x new weapons, he will pay x^2 coins for it. 
Also note that Ben can carry all the weapons he has currently to the next level . Initially, Ben has no weapons. 
Can you tell the minimum coins required such that Ben can clear all the levels.